Added actions per minuet and action state changes per minute stats to analyzer; added total counts for stick movements and button pushes; added max GALINT stat

+++ src/replay.h
@@ -31,7 +31,7 @@ struct SlippiFrame {
   float    c_x           = 0.0f;
   float    c_y           = 0.0f;
   float    trigger       = 0.0f;
-  uint32_t buttons       = 0;
+  uint16_t buttons       = 0;
   float    phys_l        = 0.0f;
   float    phys_r        = 0.0f;
   uint8_t  ucf_x         = 0;
